+++ vkworkspace/filters/__init__.py
@@ -2,6 +2,7 @@ from .base import BaseFilter
 from .callback_data import CallbackData
 from .chat_type import ChatTypeFilter
 from .command import Command, CommandObject
+from .message_parts import ForwardFilter, RegexpPartsFilter, ReplyFilter
 from .regexp import RegexpFilter
 from .state import StateFilter
 
@@ -11,6 +12,9 @@ __all__ = [
     "ChatTypeFilter",
     "Command",
     "CommandObject",
+    "ForwardFilter",
     "RegexpFilter",
+    "RegexpPartsFilter",
+    "ReplyFilter",
     "StateFilter",
 ]